requirements.txt
streamlit
python-dotenv
opencv-python
google-generativeai
Pillow